pyproject.toml
[tool.poetry]
name = "pubmed_fetcher_reasearch_paper"
version = "0.1.3"
description = "A CLI tool to fetch and summarize PubMed research papers using LLMs, embeddings, and RAG."
authors = ["Pavan Kalyan <neelampavan95@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "pubmed_fetcher", from = "src" }]
homepage = "https://github.com/Pavan-Kalyan112/Research-Paper-Fetcher-CLI"
repository = "https://github.com/Pavan-Kalyan112/Research-Paper-Fetcher-CLI"
keywords = ["pubmed", "cli", "llm", "rag", "bioinformatics", "papers", "summarization"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.25.0"
pandas = "^1.5.3"
requests = "^2.32.4"
rich = "^14.0.0"
fpdf = "^1.7.2"
lxml = "^6.0.0"
beautifulsoup4 = "^4.13.4"
python-dotenv = "^1.1.1"
sentence-transformers = "^5.0.0"
faiss-cpu = "^1.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.scripts]
pubmed-fetcher = "pubmed_fetcher.cli:main"
embed-papers = "pubmed_fetcher.embedder:main"
rag-assistant = "pubmed_fetcher.rag:main"
